Quantize pressure sensor measurements.
This includes sampling from a quantized and clipped normal distribution to
determine the probability of a measurement given the estimated state.

+++ fc.h
@@ -19,6 +19,6 @@ void tick(double delta_t);
 void arm(void);
 void launch(void);
 void accelerometer_sensor(vec3 acc);
-void pressure_sensor(double pressure);
+void pressure_sensor(unsigned pressure);
 
 #endif /* FC_H */
